Bring feedback from @dcbriccetti's review Replace the old test with the new
