+++ src/test/java/org/opendatakit/common/persistence/TestRow.java
@@ -1,3 +1,18 @@
+/*
+ * Copyright (C) 2018 Nafundi
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
 package org.opendatakit.common.persistence;
 
 import java.util.Date;
@@ -15,11 +30,11 @@ class TestRow {
     public final Boolean booleanField;
 
     TestRow(String someString, Integer someInteger, Double someDouble, String someDate, Boolean someBoolean) {
-        this.stringField = someString;
-        this.integerField = someInteger;
-        this.doubleField = (someDouble == null) ? null : WrappedBigDecimal.fromDouble(someDouble);
-        this.dateField = WebUtils.parseDate(someDate);
-        this.booleanField = someBoolean;
+        stringField = someString;
+        integerField = someInteger;
+        doubleField = (someDouble == null) ? null : WrappedBigDecimal.fromDouble(someDouble);
+        dateField = WebUtils.parseDate(someDate);
+        booleanField = someBoolean;
     }
