+++ src/test/java/org/opendatakit/common/persistence/TestTable.java
@@ -1,3 +1,18 @@
+/*
+ * Copyright (C) 2018 Nafundi
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
 package org.opendatakit.common.persistence;
 
 import static org.opendatakit.common.persistence.DataField.DataType.BOOLEAN;
@@ -6,9 +21,7 @@ import static org.opendatakit.common.persistence.DataField.DataType.DECIMAL;
 import static org.opendatakit.common.persistence.DataField.DataType.INTEGER;
 import static org.opendatakit.common.persistence.DataField.DataType.STRING;
 
-import org.opendatakit.common.persistence.exception.ODKDatastoreException;
 import org.opendatakit.common.security.User;
-import org.opendatakit.common.web.CallingContext;
 
 class TestTable extends CommonFieldsBase {
   static final DataField stringField = new DataField("STRING_FIELD", STRING, true, 90L);
